Functional Favourites page
Completed functionality of favourites page. Also adjusted icons on calendar page to have the same icons (or as close as possible) to the favourites page. Need to adjust the middle star icon of the calendar page to match the favourites page. Made the calendar page buttons at the bottom clickable, need to make the top buttons also clickable